Lock Computer
This opens the application that Apple Remote Desktop opens when locking
end users out of their computer.

The policy was set to launch on boot because all that is required to
unlock is to reboot.

Remove the user from the policy so that the next time they reboot, they
have access to their system.