Edits PlayReady for netflix
This checks to see if the file required for netflix playback is
present. If so, it temporarily copies it to a temp directory. Sets
appropriate permissions on the PlayReady Folder so students can’t hide
content inside this folder, it then moves the netflix file for playback
as now they cannot write into the PlayReady folder.

If looking to simply remove the netflix file all together. change cp in
line 7 to rm and remove code from line 13 down.